Inline dialog header styles and rename DialogTitle to DialogHeader
Remove the header* and title* slots from dialogRecipe (headerSide,
headerButton, headerSearchRow, titleContainer, titleText) and inline them
as Panda css() at their single call sites in DialogHeader, CircleButton,
and MobileCommandUniverse. The recipe now owns just the glass shell,
decorative layers, and scrollable content region.

Rename DialogTitle -> DialogHeader (component, props, file) now that it
styles itself entirely inline. No visual change — same generated CSS.
